Draft: move get_DXF to utils.py
